BUGFIX: Send reminders on behalf of direct reports
Previously, attempting to send a reminder on behalf of a direct report
would result in a forbidden response, which appears to be the logged out
page, although the user wasn't actually logged out.

I'm not entirely happy with the implementation, especially in
_send_reminder.html.erb and line 11 of reminders_controller.rb. A
cleaner solution probably requires an object that can be routed, so that
we can say `form_for [@subject, reminder].compact`. However, in the
interests of fixing the immediate problem, this helps.